+++ src/components/point-charge-modal/info-container/index.tsx
