+++ src/components/point-charge-modal/point-modal/index.tsx
