+++ src/components/point-charge-modal/point-modal/types.ts
