+++ src/components/point-detail/badge/index.tsx
