+++ src/components/point-detail/index.tsx
