+++ src/components/point-detail/payment/common/company-info/index.tsx
