+++ src/components/point-detail/payment/common/coupon-info/index.tsx
