+++ src/components/point-detail/payment/common/order-info/index.tsx
