+++ src/components/point-detail/payment/common/order-point-info/index.tsx
