+++ src/components/point-detail/payment/common/order-point-info/types.tsx
