+++ src/components/point-detail/point-box/index.tsx
