+++ src/components/point-detail/point-detail-list/index.tsx
