+++ src/components/point-detail/point-detail-menu/index.tsx
